+++ src/backend/external_canisters/src/ledger.rs
@@ -5,7 +5,7 @@ use icrc_ledger_types::icrc1::account::Account;
 pub struct LedgerCanisterService(pub Principal);
 
 impl LedgerCanisterService {
-    pub async fn icrc1_balance_of(&self, account: Account) -> CallResult<u64> {
+    pub async fn icrc1_balance_of(&self, account: Account) -> CallResult<u128> {
         let (res,) = call(self.0, "icrc1_balance_of", (account,)).await?;
         Ok(res)
     }
